Move plugin config extract to Simulation
Started plugins are kept track in the simulation,
so move the code for extracting their config
into simulation as well.
